renamed generated to js_interlop
