+++ packages/google_sign_in/google_sign_in_web/lib/src/js_interlop/gapiauth2.dart
